added Query that resolves geoambiguities
Since libpostal only returns 1 result for an input, there are instances when that 1 result needs to be expanded upon since it can mean multiple things.  For example, `Luxembourg` is a country, a region (in both Belgium and Luxembourg), and a locality in Luxembourg.  Also, `Ontario` is a city in California and a region in Canada.  This Query searches ES at all coarse layers.