reconfigured queries to actually work
the queries require the use of `function_score` for population/popularity scoring which operate on the immediate `query` child.  The `function_score.query` also supports `filter` so that the hard filters can be applied at the appropriate place.

neighbourhood-only and nothing-set fallback query expected responses have also be exported to fixture files.