moved scoring functions to `query.function_score.functions`
The scoring functions have been moved to where they can be applied to the set of all results.  Part of moving them was to break the reliance on `pelias.phrase` so population_only_function.js (and others) was added to just return the scoring function instead of the wrapping `function_score`.  The new `*_function_only` views have been exposed in index.js.