moved more-nuanced placetype queries to own `should` levels
`localadmin`, `macrocounty`, `macroregion`, and `dependency` layers are supported by WOF but not libpostal.  For instance, "Puerto Rico" is considered by libpostal to be a country:

```
> puerto rico

Result:

{
  "country": "puerto rico"
}```

even though WOF considers it to be a [dependency](https://whosonfirst.mapzen.com/spelunker/id/85633729/#2/52.3/0.3).

This PR adds top-level `should`s for these 4 layers and completes the secondary matches.