added support for `input:postcode` match when available
This is overly-restrictive since it requires the supplied post code to be correct, otherwise it falls back to the city match.  This is intended to resolve `1 main st ny 11201` so it finds the correct `1 main st` but has the side effect that `1090 n charlotte st, lancaster, pa 17601` will fall back to Lancaster, PA since the correct post code for that address is 17603.