added locality-as-borough if no borough supplied
This is meant to cover the case where a locality is supplied but a borough is not.  This should be exceedingly rare, but an example of this is when a caller supplies `city=Manhattan` and `state=NY`.  Normally this wouldn't return anything because Manhattan is actually a borough in the data but this fix says that "if no borough is supplied but locality is, treat the locality value as if it might be a borough".