feat(query): Use full form for match_phrase query
There are two forms for the Elasticsearch [match_phrase](https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-match-query-phrase.html) query:

1. A shorter, condensed version which does not allow setting extra
options:
```
{
  "match_phrase": {
    "the-field": "the search terms"
  }
}
```

2. A complete version that does allow setting options:
```
{
  "match_phrase": {
    "the-field": {
      "query": "the search terms",
      "some-option": "some-value"
    }
  }
}
```

Pelias currently uses a mix of the two formats. This change removes all
usage of the condensed format in favor of the full format.

As a pure refactoring, with no functional changes, this should be safe
to merge immediately. However, it will make it easier to add new options
to phrase queries (such as `slop`) in the future.