use bounding boxes in address searches
This change helps when people specify an address in a very nearby town
or neighborhood to the locality where it is, strictly speaking,
contained by.

It is entirely backwards compatible, and requires on a small change to
pelias/api.

It does so by modifying the current parent layer id filter query into a
should between bboxes & ids. It then makes up for this increased recall
with a should clause in the query that boosts results with exact parent
layer id matches.

One example is searching for an address in Stanford, CA but specifying
it as being in Palo Alto, like "553 Mayfield Ave, Palo Alto, CA 94304, US"
- previously this query would have entirely failed.

In my testing, this is almost 100% positive. One regression I've seen is
when there is an exact match in an adjoining locality, but an interpolated
match in the specified locality. Example: 419 26th St, Oakland, CA
94612, USA. I am open to ideas about how to tell the query
"interpolation + parent id is better than exact + bbox".